requirements.txt
tensorflow==2.18.0
scikit-learn==1.5.2
numpy==2.0.2
pillow==11.0.0
opencv-python==4.10.0
streamlit==1.40.2